[BUGFIX] Show hash to user even if InstallTool password is invalid
Currently, if a user uses an invalid InstallTool password hash,
for example a plaintext placeholder, the InstallTool login
throws an uncaught `InvalidPasswordHashException`.

This prevents the output of the generated hash based on
the entered password, which could then be used to fix the
issue by setting a valid InstallTool password in the
configuration.

With this change, the exception is caught and is written
to the log file instead. The user is shown the normal login
screen with the generated hash.

Resolves: #105016
Releases: main, 13.4
Change-Id: Ib0a3b1dafba23a4554eb11d4adbad51e7f07527c
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/86147
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Andreas Nedbal <andy@pixelde.su>
Reviewed-by: Andreas Nedbal <andy@pixelde.su>
Reviewed-by: Simon Praetorius <simon@praetorius.me>
Tested-by: Simon Praetorius <simon@praetorius.me>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>