[TASK] Make pagination context aware
We are dropping the Bootstrap CSS definitions for
all paginations to have more control about the styling
of the pagination elements. From now on we are
shipping custom css adapted to our needs. The new
pagination element styling is context aware and adapts
to light and dark color schemes.

Resolves: #105032
Releases: main
Change-Id: I5e49e87694d75dc483b0acd5f83d96297e5b8c78
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/86182
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Benjamin Franzke <ben@bnf.dev>
Tested-by: Benjamin Franzke <ben@bnf.dev>
Reviewed-by: Benni Mack <benni@typo3.org>
Reviewed-by: Andreas Kienast <akienast@scripting-base.de>
Tested-by: Andreas Kienast <akienast@scripting-base.de>