[TASK] Relocate hidden elements toggle to docheader in pagelayout
This change aligns with the ongoing effort to centralize view-related
settings in the docheader. The hidden elements toggle in the page
layout view has been moved to the docheader for a more consistent
UI. The toggle is now always visible, even when no hidden elements
are present on the page.

Resolves: #105040
Releases: main
Change-Id: Ida77282703c3b7a79381c74515c86ddc88d427b5
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/86196
Reviewed-by: Andreas Kienast <akienast@scripting-base.de>
Reviewed-by: Benjamin Franzke <ben@bnf.dev>
Tested-by: Andreas Kienast <akienast@scripting-base.de>
Tested-by: Benjamin Franzke <ben@bnf.dev>
Tested-by: core-ci <typo3@b13.com>