[TASK] Revert "[TASK] Improve debug output of PAGEVIEW"
This reverts commit de1d92aab19b5c34105374090d89db8548a7e592.

Reason for revert:
The adressed debugging usability issue (the nested evaluation of the
language object in scope of the site object in output of <f:debug>)
has been adressed by #105043.

That means we can go back to logical ordering that reflects the
order how these variables have been resolved by the middleware
stack and are therefore logically tangible:

 * The `site` is resolved based on a request
 * The `language` is resolved based on a site of the request
 * The `page` is resolved based on the `site` and `language`

This logical structure helps developers to get a comprehensible overview
of variables instead of having to grasp a random ordered pile of
variables.

Reverts: #105013
Resolves: #105045
Releases: main, 13.4
Change-Id: Ie70564afd083bbd120f291aba07a20605e5574d0
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/86214
Tested-by: core-ci <typo3@b13.com>
Tested-by: Benjamin Franzke <ben@bnf.dev>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Benjamin Franzke <ben@bnf.dev>