[DOCS] Move changelogs relevant for 12.4.x to matching directory
Two changelog ReST files were initially placed in the 13.x
structure, but were backported to 12.x.

This patch moves one of the ReST files to the `12.4.x` directory,
also to raise awareness for the next 12.4.x release so
that people who might have missed these two changelogs
can catch up. The other is irrelevant for 12.4.x and
thus removed in the backport (not in main).

Resolves: #105050
Related: #101776
Related: #104108
Releases: main, 12.4
Change-Id: I468e207353f63d5574ecb8abaf5788ed0890270f
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/86205
Reviewed-by: Garvin Hicking <gh@faktor-e.de>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: Guido Schmechel <guido.schmechel@brandung.de>
Reviewed-by: Guido Schmechel <guido.schmechel@brandung.de>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: Lina Wolf <112@linawolf.de>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Lina Wolf <112@linawolf.de>
Tested-by: Garvin Hicking <gh@faktor-e.de>