[BUGFIX] Fix additional attribute handling in MediaViewHelper
With deprecation #104223 some attributes for TagBased ViewHelpers are
available in `$this->additionalArguments` which were previously
located in `$this->arguments`. This has been addressed for Core
ViewHelpers in a previous patch. However, the custom implementation in
`MediaViewHelper` has been missed and thus does not work as expected.

This patch addresses the issue by merging both argument arrays, which
restores the previous behavior.

Also 3 new test cases are added:

* video with a title
* video with empty title
* video with title is set to "null"

Resolves: #105009
Releases: main
Change-Id: I445fdaeead368b583c19834bb76a8d23c23d6904
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/86238
Tested-by: Simon Praetorius <simon@praetorius.me>
Reviewed-by: Garvin Hicking <gh@faktor-e.de>
Tested-by: Garvin Hicking <gh@faktor-e.de>
Reviewed-by: Simon Praetorius <simon@praetorius.me>
Tested-by: core-ci <typo3@b13.com>