[TASK] Make DebuggerUtility::var_dump output context aware
DebuggerUtility::var_dump now respects the backend color
scheme (light/dark) and the color palette, if defined.

In frontend mode (when typo3 backend CSS variables are not defined)
it will render with the previous, self-defined colors as before.

To make the CSS readable/reviewable it has been refactored into array
that is compiled into a string on the fly.

Releases: main
Resolves: #105067
Change-Id: I77173b24363e6bdd62a7772d37cb4dc5087c77ff
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/86243
Tested-by: core-ci <typo3@b13.com>
Tested-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Tested-by: Andreas Kienast <akienast@scripting-base.de>
Reviewed-by: Andreas Kienast <akienast@scripting-base.de>