[BUGFIX] Properly align form labels in installation wizard
With the gradual removal of Bootstrap, form styling was only partially
adopted. As a result, "floating labels" were no longer correctly
displayed within the corresponding field, but incorrectly displayed
below it.

This patch fixes the incorrect display in the installation wizard.
Since these were the only occurrences of these "floating labels",
the corresponding styling was removed.

In the course of the work, some fields and paragraphs were rearranged
and wording was updated.

Resolves: #105124
Releases: main
Change-Id: I20a68d155c244e523af910e670e518be2eb20d4e
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/86271
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Benjamin Kott <benjamin.kott@outlook.com>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Benjamin Kott <benjamin.kott@outlook.com>