[BUGFIX] Prevent race condition while writing file caches
Whenever a cache entry is written via the cache file backends,
two race conditions could previously happen when a parallel flush
is in action:

a) The temporary file is trying to be written to a folder that is
   currently not available, because it has been moved, but not
   recreated yet. The time frame for this race condition should
   have been quite slow, but it's still the error we saw most of
   the times, which likely happened often due to stale NFS caches.

b) The temporary cache (that is created in a file named `{random}.temp`)
   could be written to the cache directory, but is (re)moved to another
   directory because a parallel `flush()` moves the containing folder to
   a temporary folder, in order to delete it afterwards.
   That means the temporary file can then still be written to (as long
   as the folder is still being deleted, the kernel file descriptor that
   PHP writes to is still valid), but the subsequent `rename` operation
   will fail because the temporary file is tried to be renamed (moved)
   from the "old" location (within the original cache folder), but is no
   longer stored there.

In order to prevent both race conditions, cache folders are no
longer removed. Only the containing cache files are deleted.

The temporary cache files are now created in a separate tmp-files
subfolder, in order to not be removed by a parallel flush operation.

Resolves: #91768
Releases: main, 13.4
Change-Id: I68883c68ca49ce9fcbe2a8816cf00059618d78e5
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/87644
Tested-by: Benjamin Franzke <ben@bnf.dev>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Benjamin Franzke <ben@bnf.dev>