[BUGFIX] Properly evaluate EXT:indexed_search restriction by mediaType
With refactorings in #102898 and #97412 the IndexSearchRepository
class property $mediaType was changed from a mixed type to
an integer, and later only integers were mapped to ENUM backed
values.

However, mediaType could always be not only an integer in range
of -2, -1 and 0 but also a string with a common file extension
like 'ppt' and so on.

This patch now allows both types (string and enum from int)
for proper search evaluation.

A functional test has been added to verify behaviour.

Resolves: #105103
Related: #102898
Related: #97412
Releases: main, 12.4
Change-Id: I0a9b9e39f0d519118ba6704a3fc35cfec4c5eaf3
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/86302
Tested-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Benni Mack <benni@typo3.org>
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Benni Mack <benni@typo3.org>