[BUGFIX] Pass correct page identifier into module state storage
With the page tree refactoring in #103043 the page tree identifier
(which contains the underscore-separated rootline) was written into
the module state storage by accident.

This identifier was reflected into the `identifier` and `selection`
module state properties.
The `identifier` was later "fixed" by implicit module state
updates in the module frame. Note: The content frames update the active
page (not the selected!) by invoking an immediate action:

  <typo3-immediate-action
    action="TYPO3.Backend.Storage.ModuleStateStorage.update"
    …

This causes `identifier` to contain the "corrected" format, but the
`selection` property was still used the wrong, underscore-separated
rootline identifier.

Since module menu makes use of the selection (and has to, in order to
retain active user selections when switching between backend modules),
this change ensures that both `identifier` and `selection` use
the same plain identifier format, as it used to before #103043.

Resolves: #103784
Related: #103043
Releases: main
Change-Id: Ic90de4faf371b2e7a93a24723c4b5c55a66193db
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/86333
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Benjamin Franzke <ben@bnf.dev>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Benjamin Franzke <ben@bnf.dev>
Tested-by: Andreas Kienast <akienast@scripting-base.de>
Tested-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Andreas Kienast <akienast@scripting-base.de>