[BUGFIX] Show/hide upload progress and info container in empty folder
Since introduction of "tile view" mode for filelist module in v12, the
progress bars were non longer visible when uploading files in an empty
folder, because the visibility of the empty parent (filelist) table
no longer toggles.

For v12, it was due to a wrong class selector
(`.t3-filelist-table-container` has been renamed to
`.t3-filelist-container`).

For main, in addition to this change, jQuery has been removed and during
this, the behaviour of the uploading files progress has changed.
Now the table that will contain the uploading files progress is created
"on the fly" and then set to hidden until upload starts.
It works fine for "tile view" but not for "list view" because this
table exists even if the folder is empty and does not have a "hidden"
attribute that serves visibility toggling.

This patch fixes these two cases separately regarding the branch :

* for v12 : just adapts the class selector
* for main : only render the table if folder is not empty

Consequently, the uploading files progress in empty folder is now
displayed in the same container for both "view" modes.
And in "list view" in a folder containing items, the progress is
displayed in the existing filelist.

This patch also fixes the hiding of the info container when the upload
starts.

Resolves: #105025
Releases: main, 12.4
Change-Id: I446d204a8d428afddff93a2f18bf9150f4631196
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/86349
Tested-by: core-ci <typo3@b13.com>
Tested-by: Garvin Hicking <gh@faktor-e.de>
Reviewed-by: Garvin Hicking <gh@faktor-e.de>