[TASK] Display errors executing the fluid:schema:generate command
Previously, the command exited with an error when an exception occured
with Fluid 4.0.0. With Fluid 4.0.1 possible errors are caught and collected.

The command now fetches these possible errors and outputs them on the console. This allows extension authors to see those problems which may
hint to technical debts.

Resolves: #105242
Related: #105187
Releases: main
Change-Id: Id0d75dd40d3f94e04fdbf76ac4d3d87f80db45a8
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/86360
Tested-by: Garvin Hicking <gh@faktor-e.de>
Tested-by: Simon Praetorius <simon@praetorius.me>
Reviewed-by: Garvin Hicking <gh@faktor-e.de>
Reviewed-by: Simon Praetorius <simon@praetorius.me>
Tested-by: core-ci <typo3@b13.com>