[BUGFIX] Fix AbstractFormViewHelper type hint for identity map
With a larger refactoring of #96473 type hinting was introduced
for `AbstractFormViewHelper->renderHiddenIdentityField()`.

The phpdoc annotation previously hinted at the input being of
type 'object' (or null), and the type hint was thus set to
'object', too.

However, the argument restriction of a ViewHelper like
`FormViewHelper` defined an 'object' of type 'mixed',
and would thus allow arrays. Arrays do not need a
hidden identity field, so returning '' for that
would allow the ViewHelper to function properly.

With this patch, the type hint is changed to 'mixed'
now, so that previous ViewHelper usages with arrays
should be restored again and not lead to a type
exception.

Resolves: #105146
Resolves: #103162
Related: #96473
Releases: main, 12.4
Change-Id: I8a431e8909efb2e8d5ad041c69a0c75b7296c721
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/86367
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Reviewed-by: Garvin Hicking <gh@faktor-e.de>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: Garvin Hicking <gh@faktor-e.de>
Tested-by: core-ci <typo3@b13.com>