[TASK] Add link from the constant editor to the site setting editor
With the introduction of the site settings editor in #104794,
we can now point the user in set-provided TypoScript records
to the respective site configuration/settings modules
in the TypoScript Edit and Constant Editor view.

Also the info labels are slightly adapted to take the settings
editor into account.

Resolves: #104897
Releases: main, 13.4
Change-Id: I25f803fab123d823fd2edc8de7c1ce3e71b3b869
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/87638
Reviewed-by: Benjamin Franzke <ben@bnf.dev>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Benjamin Franzke <ben@bnf.dev>