[TASK] Add link from the constant editor to the site setting editor
With the introduction of the site settings editor in #104794,
we can now point the user in set-provided TypoScript records
to the respective site configuration/settings modules
in the TypoScript Edit and Constant Editor view.

Also the info labels are slightly adapted to take the settings
editor into account.

Resolves: #104897
Releases: main
Change-Id: I25f803fab123d823fd2edc8de7c1ce3e71b3b869
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/86391
Reviewed-by: Benni Mack <benni@typo3.org>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Jasmina Ließmann <minapokhalo+typo3@gmail.com>
Tested-by: Benjamin Franzke <ben@bnf.dev>
Reviewed-by: Benjamin Franzke <ben@bnf.dev>
Tested-by: Jasmina Ließmann <minapokhalo+typo3@gmail.com>