[BUGFIX] Use view transitions for color-scheme and theme changes
To prevent inconsistent per-element color-transitions during a
color-scheme or theme change, the per-element transitions are now
disabled for the time-frame where the color-scheme or theme attribute
is changed and a view transition is applied now.
This instructs the browser to take a screenshot before and after
the view transition and animate between.

For browsers which do not support view transitions (firefox)
or for users that prefer reduced motion we apply the changes
immediately (with transitions disabled).

Resolves: #105177
Releases: main
Change-Id: I37e27b6ddcdc54ff3e944b5c33b2797a0b033c69
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/86397
Tested-by: Andreas Kienast <akienast@scripting-base.de>
Tested-by: Garvin Hicking <gh@faktor-e.de>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Benjamin Franzke <ben@bnf.dev>
Tested-by: Benjamin Franzke <ben@bnf.dev>
Reviewed-by: Andreas Kienast <akienast@scripting-base.de>
Reviewed-by: Garvin Hicking <gh@faktor-e.de>