[BUGFIX] Immediately apply backendTitleFormat changes in user setup
The user setup module now emits an event that is immediately
applied to the module router by the user settings manager,
without requring a full backend reload.

The module router is adapted to react on this attribute
change without reloading the entire module route (therefore
`sitenameFirst` is no longer a reactive property that'd cause a
re-rendeering), but applies the browser title update reactively.

Releases: main
Resolves: #105178
Change-Id: Ic6dc1064f5b10ed516680217c97d6e1d11f5a1f8
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/86398
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Benjamin Franzke <ben@bnf.dev>
Tested-by: Frank Nägler <frank.naegler@typo3.com>
Tested-by: Andreas Kienast <akienast@scripting-base.de>
Reviewed-by: Frank Nägler <frank.naegler@typo3.com>
Reviewed-by: Andreas Kienast <akienast@scripting-base.de>
Tested-by: Benjamin Franzke <ben@bnf.dev>