[BUGFIX] Repair core-updater result and avoid JS error
With #101576 the install tool Progressbar was migrated to lit,
and jquery code (which silently ignores non available elements)
tried to remove new progressbar component that was actually never
rendered in the core-updater – it only shows a loading message.

With the migration from jQuery to querySelector in #101655, this became
an issue, since the element is not found and `remove()` was called on
null, leading to an JS error and aborting the action. That means the
core updater result was never shown.

The loading flash message is now gracefully removed, so that the
result is shown in any case.

The loading message should be migrated to the new progress bar
component, but that's up for another change.

Releases: main
Resolves: #104395
Related: #101576
Related: #101655
Change-Id: Ib904fbe14806830cc0b53aeb5aaa7762d45db608
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/86414
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Benjamin Kott <benjamin.kott@outlook.com>
Tested-by: Benjamin Kott <benjamin.kott@outlook.com>
Reviewed-by: Andreas Kienast <akienast@scripting-base.de>
Tested-by: Andreas Kienast <akienast@scripting-base.de>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>