[BUGFIX] Fix pagetree subtree expansion in search mode
The TreeController was adapted with #104397 to reduce possibly
unneeded getTreeLevels calls and refactored the condition that
sets the hasChildren flag on a node object.
It turns out that this caused search queries to miss the
hasChildren flag, as the depth pre-condition always evaluated
to false (treeLevelsToFetch is 999 is search-mode).

Resolves: #105239
Related: #104397
Releases: main, 12.4, 11.5
Change-Id: I08dc14ae045afd8a526b83ac2b356841af84eacb
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/86535
Reviewed-by: Oliver Hader <oliver.hader@typo3.org>
Reviewed-by: Benjamin Franzke <ben@bnf.dev>
Tested-by: Benjamin Franzke <ben@bnf.dev>
Tested-by: core-ci <typo3@b13.com>