[BUGFIX] Remove test namespace for Filemetadata
With #105240 all test files of the filemetadata system
extension have been removed. Thus, the composer root now
defines a namespace for a folder that does not exist anymore.
This patch removes the obsolete definition from the composer
root file.

The underlying bug in the new Fluid schema generator,
which helped discover this issue, will be fixed with Fluid 4.0.2.

Resolves: #105260
Related: #105240
Releases: main
Change-Id: Ie9898844d0c948bc42ecaa85e971115204785a34
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/86542
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Simon Praetorius <simon@praetorius.me>
Tested-by: Simon Praetorius <simon@praetorius.me>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>