[BUGFIX] Prevent undefined array key access in 'removeTags' parsing
This patch addresses a regression introduced in the TypoScript
config parsing of the 'removeTags' processing. When 'removeTags'
was set as a string (not an array), a PHP "Undefined array
key access" warning occurred.

Tests are added to parse TSconfig variant for strings and
arrays, in addition to the existing YAML-based tests.

Resolves: #105254
Related: #86215
Releases: main, 12.4
Change-Id: Iad76662a9e09bd957a0847b30473a78fa33d195f
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/86543
Tested-by: Georg Tiefenbrunn <tig@earlybird.at>
Reviewed-by: Oliver Hader <oliver.hader@typo3.org>
Tested-by: Benjamin Franzke <ben@bnf.dev>
Tested-by: Oliver Bartsch <bo@cedev.de>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Garvin Hicking <gh@faktor-e.de>
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Tested-by: Garvin Hicking <gh@faktor-e.de>
Reviewed-by: Benjamin Franzke <ben@bnf.dev>
Reviewed-by: Georg Tiefenbrunn <tig@earlybird.at>