[TASK] Ensure corect value preparation for test assertion
This change prepares data converting it forth and back on
the retrieved value and the expectation value. Note, that
this is required as different databases persists json data
differently into the columns, and return the values slightly
reformatted.

For example, removing spaces between colons and double-quotes:

* put in:               {"key": "value"}
* saved and returned as {"key":"value"}

Avoiding "concrete" expectations for each database vendor
and version, two way convertion is used on expected and
retrieved data. The test goal is to match that the json
data could be saved at all and is not changed by this change.

This test has been recently added with #105208 but did not
failed on execution on the change directly.

[1] https://git.typo3.org/typo3/CI/cms/-/jobs/3856803

Resolves: #105280
Related: #105208
Releases: main
Change-Id: If2b16b35476927aabbb6b336ca6eb1c9ddb60332
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/86562
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Garvin Hicking <gh@faktor-e.de>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Garvin Hicking <gh@faktor-e.de>