[TASK] Merge AdminPanelDataPersister into AdminPanelRenderer
The admin panel suffers from a few issues that are
complex to track down and to get right due to the
complex interplay of the various involved admin
panel middlewares.

As a first patch in this area we simplify the
construct a bit.

The admin panel comes with four middlewares:
* The pretty early AdminPanelInitiator
* An intermediate SqlLogging (not of relevance
  with this patch)
* AdminPanelRenderer that renders the admin
  panel and merges it into response body using
  MainController
* AdminPanelDataPersister that sets gathered
  data needed for AdminPanelRenderer into
  MainController

AdminPanelDataPersister is located directly after
AdminPanelRenderer in the middleware chain. Both
middlewares do things *after* calling
$handler->handle($request), so the code executed
in AdminPanelDataPersister is run directly *before*
the code in AdminPanelRenderer.

The patch merges the code from AdminPanelDataPersister
into AdminPanelRenderer directly before the existing
handling of AdminPanelRenderer which obsoletes
AdminPanelDataPersister.

Resolves: #105291
Related: #104724
Releases: main
Change-Id: Ie3d3411b6968761c9e3f9381dd8e838e8e2b3112
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/86571
Tested-by: Garvin Hicking <gh@faktor-e.de>
Reviewed-by: Garvin Hicking <gh@faktor-e.de>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>