[TASK] Avoid adminpanel_requestcache
With #105291 it became obvious that admin panel
MainController->storeData() and render() are called
directly after each other.

storeData() until now created a cache entry to set() into
adminpanel_requestcache, just to be get() it in render()
again.

This triggered serialization of ModuleDataStorageCollection,
which is an \SplObjectStorage, leading to fails when attached
objects contain closures. It also does not play well with
injected services since it can create huge amount of data.

Additionally, adminpanel_requestcache is a database backend
and the serialized data is easily in the range of multiple
megabytes, which leads to massive INSERT and SELECT queries.

ModuleDataStorageCollection should most likely not extend
\SplObjectStorage anymore in the future, but changing this
might a be breaking change for extensions that extend
admin panel in v13.

But we can get rid of adminpanel_requestcache altogether
without being breaking: This step within MainController can
be left out easily, fully avoiding that cache. The patch does
this and fixes the crippled DI in consuming admin panels.

Persisting admin panel data within adminpanel_requestcache
was probably designed to allow extending the admin panel
with a BE module to introspect past requests. This however
never materialized. If such a functionality is added later
it should probably not rely on the core cache infrastructure
since that is a misuse and should come with a redesign of
ModuleDataStorageCollection to avoid serializing the service
objects.

Resolves: #104724
Related: #105291
Releases: main
Change-Id: I8b7111ab39ecf779310418a3e7b3d4cd2024f23a
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/86572
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Andreas Nedbal <andy@pixelde.su>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Andreas Nedbal <andy@pixelde.su>