[BUGFIX] Avoid list_type deprecation message in TCA ext_tables.php check
When TCA ext_tables.php check is executed in install tool, there is
no extension configuration. This leads to
`ExtensionManagementUtility::addPlugin()` being called with
`list_type` as type, because `ExtensionUtility::registerPlugin()`
falls back to `list_type`, if no plugin type for an extension
can be evaluated via `EXTCONF`.

This change prevents the deprecation being logged, by checking
the instance of the static package manager in
`ExtensionManagementUtility`. In case it is an instance of
`FailsafePackageManager`, the call to
`ExtensionManagementUtility::addPlugin()` is invoked via
install tool.

Resolves: #105307
Releases: main
Change-Id: Icc6cfb682d646abafed5f7ca9024ba62427f81ee
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/86588
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Tested-by: Nikita Hovratov <nikita.h@live.de>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Kevin Appelt <kevin.appelt@icloud.com>
Reviewed-by: Kevin Appelt <kevin.appelt@icloud.com>
Tested-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Nikita Hovratov <nikita.h@live.de>