[TASK] Remove unneeded checks for rendering the record list search box
Having the check `$tableListHtml || !empty($this->searchTerm` prevents
the rendering of the search box if no search term is present and also
no items exist on the current page.

When using a table filter (single table view), it could be possible
that items are available in sub pages, but since the search level
selector is bound to the search box, it is impossible to change this
level and adjust the search depth.

Releases: main, 13.4
Resolves: #105313
Related: #99895
Change-Id: I1cce5435c3f893fe0193c04147fd3b055ca08593
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/86606
Tested-by: core-ci <typo3@b13.com>
Tested-by: Garvin Hicking <gh@faktor-e.de>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Garvin Hicking <gh@faktor-e.de>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>