[BUGFIX] Adjust displayed ENABLE_INSTALL_TOOL legacy install location
The backend screen that shows the "Create an enable install tool file"
message was adapted in #104126, mainly aimed for composer-mode.

Instead of pointing to "typo3conf/" it was also adjusted to "config/".

This patch restores the displayed target directory for an
ENABLE_INSTALL_TOOL file to "typo3conf/". In composer-mode, it
will point to "var/transient/".

Resolves: #105326
Related: #104126
Related: #104074
Releases: main, 13.4
Change-Id: I3d92d76679c7a4ba6d1d95404ca0566308d10502
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/86632
Tested-by: Benjamin Franzke <ben@bnf.dev>
Reviewed-by: Benjamin Franzke <ben@bnf.dev>
Tested-by: Simon Schaufelberger <simonschaufi+typo3@gmail.com>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Andreas Kienast <akienast@scripting-base.de>
Reviewed-by: Simon Schaufelberger <simonschaufi+typo3@gmail.com>
Reviewed-by: Andreas Kienast <akienast@scripting-base.de>