Prevent "id" parameter in backend URLs if not passed
In TYPO3 v13, namespaced URL GET parameters are not supported any more and the URLs look like "/typo3/module/web/aimeos/Jsonadm/index?token=...".

If the Extbase plugin uses an "id" parameter to identify a record, this "id" parameter will be added by buildBackendUri() to all links generated for that request, even if the generated links must not include an "id" parameter, e.g. when searching in repositories.

This behavior seems to be a left-over from previous TYPO3 versions. The "id" parameter should be only set if setTargetPageUid() is used and a valid page ID (>0) is passed.