[BUGFIX] Generate correct overlay icon when hiding record in record list
When hiding a record in the record list, a placeholder icon with a
hard-coded overlay `overlay-hidden` was generated.
This is the default overlay, unless an event listener for the event
`ModifyRecordOverlayIconIdentifierEvent` is implemented that may
override this.
The overridden icon is correctly shown when initially rendering the
records, but not when visible icons get hidden.

This patch now always generates a dummy icon in a forced hidden state to
get the overlay icon identier, which is then used in the toggle action.

Resolves: #105316
Releases: main, 13.4, 12.4
Change-Id: I03527f0eda9fe891565a75fa4cb3decbb5818c47
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/86643
Reviewed-by: Garvin Hicking <gh@faktor-e.de>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Andreas Kienast <akienast@scripting-base.de>
Tested-by: Andreas Kienast <akienast@scripting-base.de>
Tested-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Tested-by: Garvin Hicking <gh@faktor-e.de>