Remove validation for "id" parameter in backend
Since backend module parameters are not namespaced any more in v12/v13, the "id" parameter can not be treated as page ID any more because extbase extensions use "id" to identify own records.

The problem occurs in v12 because there, namespaced and non-namespaced configurations are supported. In Extbase extensions using non-namespaced parameters, the "id" parameter is still validated as backend page ID even if it's an Extbase module identifier.