[BUGFIX] Cast tag-content of several ViewHelpers to string
Several Fluid ViewHelpers used the method:

```
$this->tag->setContent($this->renderChildren());
```

However, in many cases specifying an integer as a tag content (for
example in pagination cases) is valid:

```
<f:for each="{1:'step1', 2:'step2'}"
       as='i'
       iteration='iterator'
       key="k">
    <f:link.action action="{i}">{k}</f:link.action>
</f:for>
```

This would lead to an exception, because the `setContent()` method
expects a a string.

Several ViewHelpers have now been adapted to cast the result of
`renderChildren` to a string.

Tests verifying this behaviour are added. Some cases where ViewHelpers
already had string-casting in effect are now secured with a test, too.

Notable exemptions of ViewHelpers where no string casting has been
applied:

- EXT:form VHs already seem to take care of proper renderChildren()
  handling.
- f:uri.action and f:uri.page do not accept tag based content.
- ce:link.clickEnlarge, sg:example, sg:code, sg:colorscheme do
  not make sense to receive non-string input, exceptions still

Note that for <dashboard:widget.button> no functional test exists yet,
and is beyond scope of this patch to be added.

Drive-by fix of adding (string) cast to an integer-based
'title' attribute for <f:be.infobox> added as well.

Resolves: #105367
Releases: main, 13.4
Change-Id: Ic1bf4f50df2ba7abb1144a2de0f18043b1c784a1
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/87087
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Garvin Hicking <gh@faktor-e.de>
Tested-by: Garvin Hicking <gh@faktor-e.de>