[BUGFIX] Ensure same RecordIdentityMap for cloned PageLayoutContext
In BackendLayoutRenderer the PageLayoutContext is cloned in case
we are in "layout" mode and a "language" is selected. The problem is
at the time of cloning, the property "recordIdentityMap" is not yet
initialized. And when calling `getGridForPageLayoutContext`, the
original context is initialized with this property and filled with
discovered Content Elements. In order to have the same identity map
in the cloned context, the map is now instantiated in the constructor
instead of lazily in the get method.

Resolves: #105153
Related: #104367
Releases: main, 13.4
Change-Id: I1fbbf734bb58df2037028f72271fd8cdef91fc6d
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/86681
Tested-by: Oliver Bartsch <bo@cedev.de>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Benni Mack <benni@typo3.org>
(cherry picked from commit fe213cb190dadcfc5785052d0ab140fe3505ddd9)
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/86852