[DOCS] Create overview page of minor-release Changelog entries by type
After the release of a LTS version, most people do not care
in which version a deprecation occured, for example.

A new overview page shows the combined features, breaking changes,
deprecations and important notes for a major version on one
page, with a table of contents.

Additionally, ":orphan:" has been removed as it is superfluous.

(Screenshot in forge ticket)

Resolves: #105391
Releases: main, 13.4
Change-Id: I3e43a8ec1bb9522dbedd0e4824a9a2007c6857ec
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/86688
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: Garvin Hicking <gh@faktor-e.de>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Reviewed-by: Garvin Hicking <gh@faktor-e.de>
Tested-by: core-ci <typo3@b13.com>