[BUGFIX] Use platform specific doctrine/dbal schema comparators
The MySQL schema comparator is aware of the fact that a column's
character set is to be inherited from its containing table, when
`TYPO3_CONF_VARS.DB.Connections.{connection}.defaultTableOptions`
is not sets.

The TYPO3 comparator wrapper implementation effectively
disabled the upstream platform adjustments for MySQL (and sqlite,
which has an own comparator as well), which is now fixed
by using a decorator pattern instead of inheritance which does
not allow to pull in platform specific comparator code.

The table options comparison is no longer done in `compareTables`,
but applied as an additional step (as the different upstream comparator
implementations are not modifiable by inheritance).

Releases: main, 13.4
Resolves: #104781
Related: #104567
Related: #104573
Change-Id: Id688720341e4f7602d13244bd1fa5741f8c95577
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/86725
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Benjamin Franzke <ben@bnf.dev>
Tested-by: Benjamin Franzke <ben@bnf.dev>
Tested-by: Garvin Hicking <gh@faktor-e.de>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Garvin Hicking <gh@faktor-e.de>
Tested-by: core-ci <typo3@b13.com>