[BUGFIX] Validate date ranges correctly
With the changes done in #104309, FormEngine works with `Date` objects
and not with timestamps anymore. This unfortunately broke the validation
in cases `lower` and/or `upper` ranges are defined as the values still
were treated as timestamps.

The code branches in the validation now have a separate check for input
fields of type `datetimepicker` and treat the comparison differently
and works with ISO8601 for the value and for the upper/lower range
values.

Resolves: #105430
Related: #104309
Releases: main, 13.4, 12.4
Change-Id: I168d1f4380167c34a3b28613bca0159737c9e8a9
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/86836
Tested-by: Benjamin Franzke <ben@bnf.dev>
Reviewed-by: Benjamin Franzke <ben@bnf.dev>
Tested-by: core-ci <typo3@b13.com>