[BUGFIX] Sync language of inline children on copy
There are three copy related DataHandler operations, each with
suble differences when elements are copied or localized to
different localizations:

* "copy" is a "free mode" copy operation
* "copyToLanguage" is a "free mode" copy operation
* "localize" is a "connected mode" copy operation

"copy" (which uses DH copyRecord()) and "copyToLanguage" (which
uses DH localize()) are quite similar, except sorting and record
prefixing details. This distinction is unfortunate and convoluted
in general, but there is also a straight issue: When using "copy"
(typically from clipboard), copied inline children do not get the
correct sys_language_uid set.

The patch adapts behavior of inline elements to that of the parent
on copy: The language field of all copied records is updated after
the actual copying has taken place. This way not only the copied
parent but also its inline relations will have the new language value
when copied to another language.

The patch implements an according loopt to DH process_cmdmap()
together with a relatively straight code refactoring and a hint
this overall scenario should be streamlined.

Resolves: #89787
Related: #104956
Related: #75400
Related: #39054
Releases: main, 13.4
Change-Id: I0f642208610beaf4fc0611944d6ce6db66c69368
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/87689
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: core-ci <typo3@b13.com>