[BUGFIX] Keep file list display settings in sync
Due to renaming the file list module to `media_management`
in #100279 where the display settings no longer shared
between the main module and the file (link) browsers.

This is now restored by using the new identifier.

Additionally, this fixes an usage of a wrong
identifier for building a return URL to the
file list module. While aliases are in place,
the module identifier has never been "file_list"
but "file_FilelistList".

Resolves: #105470
Related: #100279
Releases: main, 13.4, 12.4
Change-Id: Ieeb6dd858e938a084e4bb700261ed3c96ac78618
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/86785
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Benjamin Franzke <ben@bnf.dev>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Benjamin Franzke <ben@bnf.dev>