[DOCS] Add changelog references to breaking changes of v14
For each breaking change, the reference to the deprecation
changelog is now added. This allows developers and integrators
to find details on the removed functionality much more easily.

This patch also adds a missing entry for issue #105486
(extbase shorthand).

Note that for some changes, no changelog entry is available and
thus the forge issue is linked.

Resolves: #105488
Related: #105486
Releases: main
Change-Id: I606a6914980af0219a399431cd1b6e0339282387
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/86804
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Garvin Hicking <gh@faktor-e.de>
Reviewed-by: Lina Wolf <112@linawolf.de>
Reviewed-by: Simon Schaufelberger <simonschaufi+typo3@gmail.com>
Tested-by: Garvin Hicking <gh@faktor-e.de>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: Lina Wolf <112@linawolf.de>
Tested-by: Simon Schaufelberger <simonschaufi+typo3@gmail.com>