[BUGFIX] Provide full raw record to content object rendering
When calling the f:cObject Viewhelper by passing
a Record object as `data`, which is the case when
building frontend templates using the `page-content`
data processor, it is now ensured that the full
raw record data (all fields) are passed to the
content object rendering.

This is required for further data processors as
they would expect `$processedData['data']` to
contain the full raw record data.

An example is the ContentBlocksDataProcessor
of ext:content_blocks, which itself calls
RecordFactory to create a transformed record,
which however requires the raw input record
to contain all necessary fields, such as
the language field.

Resolves: #105518
Releases: main, 13.4
Change-Id: Ibeb035d6050bb298238884a26d6f33b9ef1316c0
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/86832
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Benni Mack <benni@typo3.org>
Reviewed-by: Nikita Hovratov <nikita.h@live.de>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Nikita Hovratov <nikita.h@live.de>
Tested-by: Oliver Bartsch <bo@cedev.de>